requirements.txt
autopep8==1.6.0
Django==4.0.*
djangorestframework==3.12.*